@@ -1,13 +1,14 @@
 import { Observable } from "rxjs/Observable";
-import { Response } from "@angular/http"
+import { HttpErrorResponse } from "@angular/common/http"
 
 export class ErrorHandler {
-    static handleError(error: Response | any): Observable<any>{
+    static handleError(error: HttpErrorResponse | any): Observable<any>{
         let errorMessage: string
-        if(error instanceof Response){
-            errorMessage = `Erro $(error.status) ao acessar a URL $(error.url) - $(error.statusText)` 
+        if(error instanceof HttpErrorResponse){
+            const body = error.error
+            errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''} ${body}` 
         }else{
-            errorMessage = error.toString()
+            errorMessage = error.message ? error.message : error.toString()
         }
         return Observable.throw(errorMessage)
     }
