@@ -3,7 +3,7 @@ import { ShoppingCartService } from "app/restaurant-detail/shopping-cart/shoppin
 import { CartItem } from "app/restaurant-detail/shopping-cart/cart-item.model";
 import { Order } from "./order.model";
 import { Observable } from "rxjs/Observable";
-import { Http, Headers, RequestOptions } from "@angular/http";
+import { HttpClient } from "@angular/common/http";
 import { MEAT_API } from "app/app.api";
 
 @Injectable()
@@ -11,7 +11,7 @@ export class OrderService {
 
     constructor(
         private cartService: ShoppingCartService,
-        private http: Http
+        private http: HttpClient
     ){}
 
     increaseQty(item: CartItem){
@@ -31,14 +31,7 @@ export class OrderService {
     }
 
     checkOrder(order: Order): Observable<string> {
-        let headers = new Headers()
-        headers.append('Content-type', 'application/json')
-        return  this.http.post(`${MEAT_API}/orders`,
-                                JSON.stringify(order),
-                                new RequestOptions({headers: headers})
-                            )
-                            .map(response => response.json())
-                            .map(order => order.id)
+        return  this.http.post<Order>(`${MEAT_API}/orders`, order).map(order => order.id)
     }
 
     clear(){
